# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"TrafficMind Agent API","description":"面向智慧交通的事件研判与闭环处置 Agent - 第一阶段 MVP","version":"1.0.0"},"paths":{"/analyze_event":{"post":{"summary":"分析交通事件","description":"输入一条交通事件 JSON，Agent 自动完成：\n  1. 事件解析与标准化\n  2. 风险评分与等级判定\n  3. 预案匹配\n  4. 处置建议生成\n  5. 调度话术生成\n  6. 结构化报告生成\n  7. 结果持久化保存\n\n返回完整的分析结果 JSON。","operationId":"analyze_event_analyze_event_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnalyzeEventRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/history":{"get":{"summary":"查询历史事件记录","description":"查询 SQLite 中保存的历史事件分析记录。\n按更新时间倒序排列，默认返回最近 50 条。\n\nArgs:\n    limit: 最大返回条数（默认50）","operationId":"history_history_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":50,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/event/{event_id}":{"get":{"summary":"查询单条事件详情","description":"根据 eventId 查询单条事件的完整分析详情。\n\nArgs:\n    event_id: 事件编号（如 E202606290001）","operationId":"get_event_event__event_id__get","parameters":[{"name":"event_id","in":"path","required":true,"schema":{"type":"string","title":"Event Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/event/{event_id}/status":{"post":{"summary":"更新事件状态","description":"更新事件处理状态。\n\n有效状态值：\n    - 待研判\n    - 待派单\n    - 处置中\n    - 已处置\n    - 待复盘\n    - 已归档\n\nArgs:\n    event_id: 事件编号\n    body: { \"status\": \"处置中\" }","operationId":"update_status_event__event_id__status_post","parameters":[{"name":"event_id","in":"path","required":true,"schema":{"type":"string","title":"Event Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StatusUpdateRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"健康检查","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/stats":{"get":{"summary":"仪表盘统计数据","description":"返回仪表盘所需的聚合统计数据：\n  - 总事件数 / 高风险事件数 / 平均风险分 / 待派单数\n  - 风险等级分布 / 事件类型分布 / 状态分布\n  - 近 7 天每日事件趋势","operationId":"stats_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AnalyzeEventRequest":{"properties":{"eventId":{"type":"string","title":"Eventid"},"eventType":{"type":"string","title":"Eventtype"},"cameraId":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cameraid","default":""},"roadName":{"type":"string","title":"Roadname"},"direction":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Direction","default":""},"lane":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Lane","default":""},"avgSpeed":{"type":"number","title":"Avgspeed"},"queueLength":{"type":"number","title":"Queuelength"},"duration":{"type":"number","title":"Duration"},"vehicleCount":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Vehiclecount","default":0},"weather":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Weather","default":"clear"},"timePeriod":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Timeperiod","default":"off_peak"},"isMainRoad":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Ismainroad","default":false},"nearbySchool":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Nearbyschool","default":false},"nearbyHospital":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Nearbyhospital","default":false},"confidence":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Confidence","default":0.9}},"additionalProperties":true,"type":"object","required":["eventId","eventType","roadName","avgSpeed","queueLength","duration"],"title":"AnalyzeEventRequest","description":"分析事件请求体（宽松模式，额外字段会被保留）"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"StatusUpdateRequest":{"properties":{"status":{"type":"string","title":"Status"}},"type":"object","required":["status"],"title":"StatusUpdateRequest","description":"状态更新请求体"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}